extract to chia.util.streamablemetadata
